+++ components-ui/app/src/tol-ui/src/interfaces/index.ts
@@ -18,6 +18,7 @@ export * from "./Forms";
 export * from "./General";
 export * from "./GlobalLoading";
 export * from "./JsonApi";
+export * from "./JsonEdit";
 export * from "./Markdown";
 export * from "./Messaging";
 export * from "./Nav";